requirements.txt
discord.py
colorama
coloredlogs